Add a fudge factor to the final summary, as used in Hesiod.
Again, the final report shows cells as 'in qc' because we don't set the
status until the final message to RT is confirmed.
Here it's fairly simple as only one pipeline instance can be active on
a run at once then at the point of summary all 'in qc' cells are actually done.
If there had been a failure on any cell then we never reach this point.