Duplex read estimation is back :-)
I'm not sure the calculation of % passing reads is right, especially given
that the duplex pass score is fixed to 9.0 but the basecall passing score
could be different. But it will do for now. And with Doarado we should (!)
just get duplex basecalling on-instrument and then we can get rid of
this bit anyway.